requirements.txt
appdirs==1.4.4
astor==0.8.1
bokeh==2.2.3
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
colorcet==2.0.2
cycler==0.10.0
Cython==0.29.21
dask==2.30.0
datashader==0.11.1
datashape==0.5.2
decorator==4.4.2
distributed==2.30.1
fsspec==0.8.4
funcparserlib==0.3.6
gnuplotlib==0.33
google-api-core==1.23.0
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-cloud-bigquery==2.2.0
google-cloud-bigquery-storage==2.0.1
google-cloud-core==1.4.3
google-cloud-storage==1.32.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
HeapDict==1.0.1
holoviews==1.13.5
hy==0.19.0
idna==2.10
imageio==2.9.0
img2unicode==0.1a7
importlib-metadata==2.0.0
jedhy==1
Jinja2==2.11.2
joblib==0.17.0
kiwisolver==1.3.1
libcst==0.3.13
llvmlite==0.34.0
locket==0.2.0
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
matplotlib-terminal==0.1a2
msgpack==1.0.0
multipledispatch==0.6.0
mypy-extensions==0.4.3
n2==0.1.7
networkx==2.5
numba==0.51.2
numpy==1.19.4
numpysane==0.28
oauthlib==3.1.0
packaging==20.4
pandas==1.1.4
pandas-gbq==0.14.0
panel==0.10.1
param==1.10.0
partd==1.1.0
patsy==0.5.1
Pillow==8.0.1
plotext==1.0.10
proto-plus==1.11.0
protobuf==3.13.0
psutil==5.7.3
pyarrow==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyct==0.4.6
pydata-google-auth==1.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
pyviz-comms==0.7.6
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rply==0.7.7
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
sortedcontainers==2.3.0
statsmodels==0.12.1
tblib==1.7.0
termplotlib==0.3.2
threadpoolctl==2.1.0
tifffile==2020.10.1
toolz==0.11.1
tornado==6.1
tqdm==4.51.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
umap-learn==0.4.6
urllib3==1.25.11
urwid==2.1.2
xarray==0.16.1
zict==2.0.0
zipp==3.4.0
